[FIX] MRIQC crashed generating csv files
when the json files for a modality was missing.
This caused that, in datasets with only (for example) anatomical images,
the group level report always failed when trying to generate the
functional group-report.

Now, MRIQC will only issue a warning when there are no measure files
for a modality.

Fixes #282