ENH: Copy NIfTI image header when creating from image
Copy the NIfTI image header of the primary source when creating another
one from the former using the `nilearn.threshold_img` function.

Make it conditional on the `nilearn` version as the warning is only
being raised for `nilearn>=0.11`.

Fixes:
```
nireports/tests/test_interfaces_segmentation.py::test_SimpleShowMaskRPT
  /home/runner/work/nireports/nireports/nireports/reportlets/mosaic.py:92:
   FutureWarning: From release 0.13.0 onwards, this function will, by
   default, copy the header of the input image to the output. Currently,
   the header is reset to the default Nifti1Header. To suppress this
    warning and use the new behavior, set `copy_header=True`.
    bbox_nii: SpatialImage = nlimage.threshold_img(bbox_nii, 1e-3)  # type: ignore[no-redef]
```

raised for exmaple in:
https://github.com/nipreps/nireports/actions/runs/12681153218/job/35344304375#step:12:355