ENH: Make use of the `compress` option in melodic plot function
Make use of the `compress` option in melodic plot function.

Since the dependencies required to perform the compression are optional,
mark the corresponding test as such.

Take advantage of the commit to streamline the search of the compression
libraries in order to reduce boilerplate code.

Add a step to the package testing GH CI workflow file to install the
required compression libraries so that the test can be run.

Co-authored-by: Mathias Goncalves <goncalves.mathias@gmail.com>