MNT: Update ruff parameters
Use extend-select to build upon the ruff default rules.

By default, ruff selects `F` rules and `E` rules that are compatible
with formatters. As a result, we don't need to explictely disable `E`
rules that are incompatible with formatters.

According to the Scientific Python Library Development Guide, pycodestyle
warnings `W` are not required when using a formatter:
	https://learn.scientific-python.org/development/guides/style/#ruff